A2SG-112 test: add initial test for staffdetails
