--- client/src/service/UpdateRoleListing.service.js
@@ -1,8 +1,0 @@
-import axios from 'axios'
-
-const BASE_URL = 'http://localhost:8080/api/v1/'
-
-export const updateRoleListing = async (updatedRoleListing) => {
-  const response = await axios.put(`${BASE_URL}rolelistings/`, updatedRoleListing)
-  return response.data
-}
